refactor(lux-offline): rename and move components and service
